working except for virtual output tests

+++ key_stack.h
@@ -7,11 +7,11 @@
 
 typedef struct
 {
-    uint16_t    *buffer;
+    uint8_t    *buffer;
     int         capacity;
-    int         size;	
+    int         size;
 } st_key_stack_t;
 
-void st_key_stack_push(st_key_stack_t *s, uint16_t key);
+void st_key_stack_push(st_key_stack_t *s, uint8_t key);
 void st_key_stack_pop(st_key_stack_t *s);
 void st_key_stack_to_str(const st_key_stack_t *s, char *str);
